added collections skeleton for loading states
